requirements.txt
requests
configparser
sqlalchemy>=1.4.36
psycopg2-binary
pandas
numpy
tushare
asyncpg
asyncio
streamlit
openpyxl
pysnowball
bs4
xlrd




